Refactoring and corrections of NucleotidSequences

+++ trunk/melting-java/src/melting/InternalLoopMethod/Turner06InternalLoop.java
@@ -49,8 +49,8 @@ public class Turner06InternalLoop extends PartialCalcul{
 
 		String mismatch1 = NucleotidSequences.getLoopFistMismatch(internalLoop.getSequence());
 		String mismatch2 = NucleotidSequences.getLoopFistMismatch(internalLoop.getComplementary());
-		int numberAU = internalLoop.calculateNumberOfTerminal('A', 'U');
-		int numberGU = internalLoop.calculateNumberOfTerminal('G', 'U');
+		double numberAU = internalLoop.calculateNumberOfTerminal('A', 'U');
+		double numberGU = internalLoop.calculateNumberOfTerminal('G', 'U');
 		
 		if (loopType.charAt(0) == '1' && Integer.parseInt(loopType.substring(2, 3)) > 2){
