+++ trunk/melting-java/src/melting/approximativeMethods/Santalucia98.java
@@ -17,7 +17,7 @@ public class Santalucia98 extends ApproximativeMode{
 	@Override
 	public ThermoResult calculateThermodynamics() {
 		double Tm = super.calculateThermodynamics().getTm(); 
-		Tm = 77.1 + 11.7 * Math.log10(this.environment.getNa()) + 0.41 * this.environment.getSequences().calculatePercentGC() - 528.0 / this.environment.getSequences().getDuplexLength();
+		Tm = 77.1 + 11.7 * Math.log10(this.environment.getNa()) + 0.41 * this.environment.getSequences().calculatePercentGC() - 528.0 / (double)this.environment.getSequences().getDuplexLength();
 		
 		this.environment.setResult(Tm);
