+++ trunk/melting-java/src/melting/cricksNNMethods/Turner06.java
@@ -81,7 +81,7 @@ public class Turner06 extends CricksNNMethod {
 		if (withoutTerminalUnpairedNucleotides == null){
 			throw new SequenceException("The two sequences can't be hybridized.");
 		}
-		int numberTerminalAU = withoutTerminalUnpairedNucleotides.calculateNumberOfTerminal('A', 'U');
+		double numberTerminalAU = withoutTerminalUnpairedNucleotides.calculateNumberOfTerminal('A', 'U');
 		double enthalpy = 0.0;
 		double entropy = 0.0;
